added authorization

+++ app/Http/Controllers/Backend/CommunityPostController.php
@@ -10,9 +10,11 @@ use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Redirect;
 use Inertia\Inertia;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 
 class CommunityPostController extends Controller
 {
+    use AuthorizesRequests;
     public function create(Community $community){
         return Inertia::render('Communities/Posts/Create',compact('community'));
     }
@@ -29,16 +31,19 @@ class CommunityPostController extends Controller
     }
 
     public function edit(Community $community, Post $post){
+        $this->authorize('update',$post);
         return Inertia::render('Communities/Posts/Edit',compact('community', 'post'));
     }
 
     public function update(StorePostRequest $request, Community $community, Post $post){
+        $this->authorize('update',$post);
         $post->update($request->validated());
 
         return Redirect::route('frontend.communities.posts.show',[$community->slug, $post->slug]);
     }
 
     public function destroy(Community $community, Post $post){
+        $this->authorize('delete',$post);
         $post->delete();
 
         return Redirect::route('frontend.communities.show',$community->slug);
